+++ js/esm-outro.js
@@ -1,0 +1,5 @@
+
+}(Bootstrap));
+
+export default Selectpicker;
+export { Selectpicker };