+++ js/i18n/defaults-pl_PL.js
@@ -3,14 +3,12 @@
  * Locale: PL (Polish)
  * Region: EU (Europe)
  */
-(function ($) {
-  $.fn.selectpicker.defaults = {
-    noneSelectedText: 'Nic nie zaznaczono',
-    noneResultsText: 'Brak wyników wyszukiwania {0}',
-    countSelectedText: 'Zaznaczono {0} z {1}',
-    maxOptionsText: ['Osiągnięto limit ({n} {var} max)', 'Limit grupy osiągnięty ({n} {var} max)', ['elementy', 'element']],
-    selectAllText: 'Zaznacz wszystkie',
-    deselectAllText: 'Odznacz wszystkie',
-    multipleSeparator: ', '
-  };
-})(jQuery);
+Selectpicker.setDefaults({
+  noneSelectedText: 'Nic nie zaznaczono',
+  noneResultsText: 'Brak wyników wyszukiwania {0}',
+  countSelectedText: 'Zaznaczono {0} z {1}',
+  maxOptionsText: ['Osiągnięto limit ({n} {var} max)', 'Limit grupy osiągnięty ({n} {var} max)', ['elementy', 'element']],
+  selectAllText: 'Zaznacz wszystkie',
+  deselectAllText: 'Odznacz wszystkie',
+  multipleSeparator: ', '
+});